Directory Module: Staff, Student and Guardian module | Status:Done

+++ plugins/Directory/src/Controller/DirectoriesController.php
@@ -610,7 +610,9 @@ class DirectoriesController extends AppController
             $session->delete('Directory.Directories.studentToGuardian');
         }
         if (($action == 'Directories') && ($this->ControllerAction->paramsPass()[0] == 'view') || empty($this->ControllerAction->paramsPass())) {
-            $session->delete('Directory.Directories.guardianToStudent');
+            if($action == 'Directories') {
+                $session->delete('Directory.Directories.guardianToStudent');
+            }
             $session->delete('Directory.Directories.studentToGuardian');
         }
 
@@ -644,9 +646,20 @@ class DirectoriesController extends AppController
             $userId = $session->read('Directory.Directories.id');
             $userId = $session->read('Directory.Directories.primaryKey.id');
 
-            if ($session->check('Directory.Directories.name')) {
+            if (isset($this->request->getParam('pass')[0]) && ($this->request->getParam('pass')[0] == 'view' || $this->request->getParam('pass')[0] == 'edit')) {
+                $param = $this->ControllerAction->paramsDecode($this->request->getParam('pass')[1]);
+                $id = $param['id'];
+                if(isset($param['staff_id']) && !empty($param['staff_id'])) {
+                    $id = $param['staff_id'];
+                } else if(isset($param['security_user_id']) && !empty($param['security_user_id'])) {
+                    $id = $param['security_user_id'];
+                }
+                $Directories = TableRegistry::getTableLocator()->get('Directory.Directories');
+                $entity = $Directories->get($id);
+                $header = $entity->name;
+            } else if ($session->check('Directory.Directories.name')) {
                 $header = $session->read('Directory.Directories.name');
-            }
+            } 
 
             $alias =$this->request->getParam('action');//$model->getAlias();
             if($alias == 'ComponentAction') {
@@ -864,11 +877,13 @@ class DirectoriesController extends AppController
 
         if (array_key_exists('userRole', $options) && $options['userRole'] == 'Guardians' && array_key_exists('entity', $options)) {
             $session = $this->request->getSession();
+            $id = (array_key_exists('id', $options))? $options['id']:$id;
             $session->write('Guardian.Guardians.name', $options['entity']->user->name);
             $session->write('Guardian.Guardians.id', $options['entity']->user->id);
             $session->write('Directory.Directories.studentToGuardian', 'studentToGuardian');
         } elseif (array_key_exists('userRole', $options) && $options['userRole'] == 'Students' && array_key_exists('entity', $options)) {
             $session = $this->request->getSession();
+            $id = (array_key_exists('id', $options))? $options['id'] : $id;
             $session->write('Student.Students.name', $options['entity']->user->name);
             $session->write('Student.Students.id', $options['entity']->user->id);
             $session->write('Directory.Directories.guardianToStudent', 'guardianToStudent');
@@ -931,7 +946,9 @@ class DirectoriesController extends AppController
             unset($tabElements[$this->getName()]);
         } elseif (array_key_exists('userRole', $options) && $options['userRole'] == 'Students') {
             $session = $this->request->getSession();
+            $id = (array_key_exists('id', $options))? $options['id'] : $id;
             $StudentGuardianId = $session->read('Student.Guardians.primaryKey')['id'];
+            $session->write('Directory.Directories.guardianToStudent', 'guardianToStudent');
             $relationTabElements = [
                 'Students' => ['text' => __('Relation')],
                 'StudentUser' => ['text' => __('Overview')]
@@ -941,6 +958,22 @@ class DirectoriesController extends AppController
             $relationTabElements['StudentUser']['url'] = array_merge($url, ['action' => 'GuardianStudentUser', 'view', $this->paramsEncode(['id' => $id, 'StudentGuardians.id' => $StudentGuardianId])]);
             $tabElements = array_merge($relationTabElements, $tabElements);
             unset($tabElements[$this->getName()]);
+            
+            foreach ($tabElements as $key => $value) {
+                if($key == 'Students') {
+                    $StudentId = $session->read('Student.Guardians.primaryKey')['id'];
+                    $security_user_id = $session->read('Student.Guardians.primaryKey')['security_user_id'];
+                    $tabElements[$key]['url'][1] = $this->ControllerAction->paramsEncode(['id' => $StudentId,'security_user_id'=> $security_user_id, 'userRole'=>'Students']);     
+                }
+                if ($key == 'StudentUser' || $key == 'Accounts') {
+                    $tabElements[$key]['url'][1] = $this->ControllerAction->paramsEncode(['id' => $id,'userRole'=>'Students']);
+                } else {
+                    $url = $tabElements[$key]['url'];
+                    $tabElements[$key]['url'] = $this->ControllerAction->setQueryString($url,
+                                                    ['security_user_id' => $id,'userRole'=>'Students']
+                                                );
+                }
+            }
         }
 
         return $this->TabPermission->checkTabPermission($tabElements);
